added full working unit tests for l1, l1

+++ CPP/l1/include/sieve.hpp
@@ -7,7 +7,7 @@ class Sieve
 {
 public:
 
-    static std::vector<size_t> primes_sieve(const size_t n) noexcept(false);
+    static std::vector<long int> primes_sieve(const long int n) noexcept(false);
 
 };
